+++ libraries/AD7367/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2025-02-04
+- add **getLastADCA()** and **getLastADCB()**
+- obsolete **getValue()**
+- updated examples (with above change).
+- add support for **ADDR** pin
+- add support for **REFSEL** pin
+- add support for **RANGE** pins
+- add examples 
+- update readme.md
+
+
 ## [0.1.0] - 2025-01-10
 - initial version